==> Data Preparation/03_Combine_VRS_Realtime.ipynb <==
# 03: Combine VRS Realtime Data Part 1

In this notebook, the VRS realtime data will be merged, aggregated and cleaned. So far, the data is seperated into thousands of single csv files.

Note: This notebooks the the data from the vrs.zip file where the realtime data come in multiple of thousands of csv files. 

Note: this notebooks was executed on local machine

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os

# Here: set the month for which you want to aggregate the data
data_month_single = '7'
data_month_double = '07'

# get a list with all files for a day

def file_names_for_each_day(day):

    if day < 10:
        day = '0' + str(day)

    folder_path = '/Volumes/T7/Master/Raw Data/' + data_month_double + '/VRS Realtime Data/'  # Replace with the actual folder path
    substring = '2023'+ data_month_double + str(day)  # Replace with the specific substring

    file_names = []

    for file_name in os.listdir(folder_path):
        if substring in file_name:
            file_names.append(file_name)

    return file_names


data = pd.DataFrame({}, columns=range(1, 32))

# create a dataframe with all files for each day in a month
for i in range(1,32):
    file_list = file_names_for_each_day(i)
    file_df = pd.DataFrame(file_list, columns=[i])
    data[i] = file_df
    

data = data.drop(data.index[1::2])

data.reset_index(drop=True, inplace=True)


def convert_times(dataframe): # convert values of stop_arrival_time and stop_departure_time to datetime
    for index, row in dataframe.iterrows():

        if row['stop_arrival_time'] == 0:
            dataframe.loc[index, 'stop_arrival_time'] = 0
        else:
            dataframe.loc[index, 'stop_arrival_time'] = datetime.fromtimestamp(int(row['stop_arrival_time']))

        if row['stop_departure_time'] == 0:
            dataframe.loc[index, 'stop_departure_time'] = 0
        else:
            dataframe.loc[index, 'stop_departure_time'] = datetime.fromtimestamp(int(row['stop_departure_time']))
    
    return dataframe

# combine the different csv files for each day
def merge_day():

    month_df = pd.DataFrame()

    for i in range(1, 32): # start with 1, change if kernel dies to last day before kernel dies
        day_df = pd.DataFrame()

        for j in range(0, len(data[i])-1):
         
            if pd.isna(data[i][j]):
                continue
                
            data_string = str(data[i][j])
         
            temp_df = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/Raw Data/"+ data_month_double + "/VRS Realtime Data/"+data_string))

            converted_data = convert_times(temp_df)

            converted_data['filename'] = data_string

            day_df = pd.concat([day_df, converted_data])

        day_df.to_csv('/Volumes/T7/Master/Processed Data/'+ data_month_double + '/day_df_'+str(i)+'.csv', index=False)

        month_df = pd.concat([month_df, day_df]) 

    month_df.to_csv('/Volumes/T7/Master/Processed Data/'+ data_month_double + '/month_df.csv', index=False) 
        

merge_day()

# load each day dataframe
# can only load 10 days at a time, otherwise kernel dies
one = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_1.csv')) 
two = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_2.csv')) 
three = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_3.csv'))
four = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_4.csv'))
five = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_5.csv'))
six = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_6.csv'))
seven = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_7.csv'))
eight = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_8.csv'))
nine = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_9.csv'))
ten = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_10.csv'))

# now concat the dataframes:

df = pd.concat([one, two])

df = pd.concat([df, three])

df = pd.concat([df, four])

df = pd.concat([df, five])

df = pd.concat([df, six])

df = pd.concat([df, seven])

df = pd.concat([df, eight])

df = pd.concat([df, nine])

df = pd.concat([df, ten])

# export the dataframe to a csv file
# need to be done in chunks, otherwise kernel dies
df.to_csv('/Volumes/T7/Master/Processed Data/'+ data_month_double + '/vrs_real_1.csv', index=False)

eleven = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_11.csv')) 
twelve = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_12.csv'))
thirteen = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_13.csv'))
fourteen = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_14.csv'))
fifteen = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_15.csv'))
sixteen = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_16.csv'))
seventeen = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_17.csv'))
eighteen = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_18.csv'))
nineteen = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_19.csv'))
twenty = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_20.csv'))

df1 = pd.concat([eleven, twelve])

df1 = pd.concat([df1, thirteen])

df1 = pd.concat([df1, fourteen])

df1 = pd.concat([df1, fifteen])

df1 = pd.concat([df1, sixteen])

df1 = pd.concat([df1, seventeen])

df1 = pd.concat([df1, eighteen])

df1 = pd.concat([df1, nineteen])

df1 = pd.concat([df1, twenty])

df1.to_csv('/Volumes/T7/Master/Processed Data/'+ data_month_double + '/vrs_real_2.csv', index=False)

twentyone = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_21.csv')) 
twentytwo = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_22.csv'))
twentythree = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_23.csv'))
twentyfour = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_24.csv'))
twentyfive = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_25.csv'))
twentysix = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_26.csv'))
twentyseven = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_27.csv'))
twentyeight = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_28.csv'))
twentynine = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_29.csv'))
thirty = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_30.csv'))
thirtyone = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/' + data_month_double + '/day_df_31.csv'))

df2 = pd.concat([twentyone, twentytwo])

df2 = pd.concat([df2, twentythree])

df2 = pd.concat([df2, twentyfour])

df2 = pd.concat([df2, twentyfive])

df2 = pd.concat([df2, twentysix])

df2 = pd.concat([df2, twentyseven])

df2 = pd.concat([df2, twentyeight])

df2 = pd.concat([df2, twentynine])

df2 = pd.concat([df2, thirty])

df2 = pd.concat([df2, thirtyone])

df2.to_csv('/Volumes/T7/Master/Processed Data/'+ data_month_double + '/vrs_real_3.csv', index=False)
