==> Data Preparation/04_Merge_Planed_with_Bonn_stations.ipynb <==
# 04: Merge first VRS "planned" data with stations in Bonn

Data that will be aggreagated, cleaned and merged in this notebook:
- Bonn Stations (Notebook 1)
- Agency
- Routes
- Stops
- Stop Times


### Note: this notebooks was executed on local machine

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
from shapely.geometry import Point
import hashlib

# Set month
data_month_single = '7'
data_month_double = '07'
data_month_int = 7

# load agency data
agency = pd.DataFrame(pd.read_csv("../Data/2023"+ data_month_double+"01-000000-UTC_gtfs-scraper_agency-vrs.csv"))

#calendar_dates = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/Raw Data/"+ data_month_double+"/VRS Planned Data/2023"+ data_month_double+"01-000000-UTC_gtfs-scraper_calendar_dates-vrs.csv"))

#calendar = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/Raw Data/"+ data_month_double+"/VRS Planned Data/2023"+ data_month_double+"01-000000-UTC_gtfs-scraper_calendar-vrs.csv"))


## Routes

# load routes data
routes = pd.DataFrame(pd.read_csv("../Data/2023"+ data_month_double+"01-000000-UTC_gtfs-scraper_routes-vrs.csv"))
routes

# if short name is missing, fill with long name
routes['route_short_name'] = routes['route_short_name'].fillna(routes['route_long_name'])

# drop columns
routes.drop(columns=['route_long_name', 'route_color', 'route_text_color'], inplace=True)

routes

# create column with route type name for better understanding
routes['route_type_name'] = routes['route_type'].apply(lambda x: 'Tram' if x == 0 else 'Rail' if x == 2 else 'Bus')

## Merge Agency with Routes

routes = routes.merge(agency[['agency_id', 'agency_name']], on='agency_id', how='left')
routes

# export to csv
routes.to_csv('../Data/04_routes_merged.csv', index=False)

Now, agency dataframe can be ignored

## Stops

# load stops data
stops = pd.DataFrame(pd.read_csv("../Data/2023"+data_month_double+"01-000000-UTC_gtfs-scraper_stops-vrs.csv"))
stops

#drop location_type
stops.drop(columns=['location_type'], inplace=True)

Now we need to narrow stops down with the stops that are only in Bonn


#load bonn stops
bonn_stops = gpd.read_parquet('../Data/01_bonn_stop_buffer.parquet')
bonn_stops = gpd.GeoDataFrame(bonn_stops)

# only keep stops in Bonn
stops = stops[stops['stop_name'].str.split().str[0] == 'Bonn']

stops

# rename columns to prepare for merge
stops.rename(columns={'stop_lon': 'long'}, inplace=True)
stops.rename(columns={'stop_lat': 'lat'}, inplace=True)

stops

# Create a new column "geometry" with Point objects
stops['geometry'] = stops.apply(lambda row: Point(row['long'], row['lat']), axis=1)

stops

# iterate over all stops and check if they are in buffer_zone
for index, stop_row in bonn_stops.iterrows():

    for index, stop in stops.iterrows():

        point = stop['geometry']

        if stop_row['buffer_zone'].contains(point):
           bonn_stops['stop_id'] = stops['stop_id']
           bonn_stops['stop_name'] = stops['stop_name']


bonn_stops

# drop rows with missing values - those are not in Bonn
bonn_stops.dropna(subset=['stop_id', 'stop_name'], inplace=True)

bonn_stops

# convert to int
bonn_stops['stop_id'] = bonn_stops['stop_id'].astype(int)

bonn_stops

# export to parquet
bonn_stops.to_parquet('../Data/04_bonn_stations.parquet') 

## Cut "Stop Times" and megre with stop data from above

Trim Stop Times to stops only in Bonn

# load stop times data
stop_times =pd.DataFrame(pd.read_csv("../Data/2023"+ data_month_double+"01-000000-UTC_gtfs-scraper_stop_times-vrs.csv"))
stop_times

# drop columns
stop_times.drop(columns=['pickup_type', 'drop_off_type'], inplace=True)

stop_times


# Merge the stop_times and merged_stops DataFrames based on the stop_id column
merged_data = pd.merge(stop_times, bonn_stops, on='stop_id', how='inner')

# The merged_data DataFrame will now contain only the rows where the stop_id matches in both DataFrames

merged_data

# create a GeoDataFrame
merged_data = gpd.GeoDataFrame(merged_data)

merged_data.info()

# export the merged data to parquet
merged_data.to_parquet('../Data/04_merged_stop_times_buffer_zones.parquet')
