==> Data Preparation/05_Reduce_VRS_to_Bonn.ipynb <==
# 05: Reduce VRS realtime data to Bonn area and merge realtime datasets

We now need to slim down the VRS realtime data to data that only contains PT trips in at our Bonn stations. After filtering, we need to combine the three realtime datasets into a single one.

### Note: This notebooks was executed on local machine 

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
import hashlib

# Set month
data_month_single = '7'
data_month_double = '07'
data_month_int = 7

## Realtime Data

# read Bonn stops data from last notebook
filtered_stops = gpd.read_parquet('../Data/04_bonn_stations.parquet')

realtime_first = pd.DataFrame(pd.read_csv("../Data/vrs_real_1.csv", on_bad_lines='skip'))


# filter to Bonn stops
filtered_realtime_first = realtime_first[realtime_first['stop_id'].isin(filtered_stops['stop_id'])]

# export to csv
filtered_realtime_first.to_csv('../Data/vrs_real_1_filtered.csv', index=False)

# load second part
realtime_second = pd.DataFrame(pd.read_csv("../Data/vrs_real_2.csv"))

#filter to Bonn stops
filtered_realtime_second = realtime_second[realtime_second['stop_id'].isin(filtered_stops['stop_id'])]

#export to csv
filtered_realtime_second.to_csv('../Data/vrs_real_2_filtered.csv', index=False)

# import third part
realtime_third = pd.DataFrame(pd.read_csv("../Data/vrs_real_3.csv"))

# filter to Bonn stops
filtered_realtime_third = realtime_third[realtime_third['stop_id'].isin(filtered_stops['stop_id'])]


# export to csv
#filtered_realtime_third.to_csv('/Volumes/T7/Master/Processed Data/'+ data_month_double + '/vrs_real_3_filtered.csv', index=False)

# need to reload if kernel dies
filtered_realtime_first = pd.DataFrame(pd.read_csv('../Data/vrs_real_1_filtered.csv'))
filtered_realtime_second = pd.DataFrame(pd.read_csv('../Data/vrs_real_2_filtered.csv'))
#filtered_realtime_third = pd.DataFrame(pd.read_csv('/Volumes/T7/Master/Processed Data/'+ data_month_double + '/vrs_real_3_filtered.csv'), engine = 'python', on_bad_lines='skip', sep=",", error_bad_lines=False)

#concatenate all files
realtime = pd.concat([filtered_realtime_first, filtered_realtime_second, filtered_realtime_third])

#export to csv
realtime.to_csv('../Data/05_vrs_all_realtime.csv', index=False)
