==> Data Preparation/06_Get_Realtime_Timestamp.ipynb <==
# 06: Get realtime timestamp

We later need to check, which was the latest status of the PT rides, to get the most accurate arrival and departure information. To do so, we need to extract the timestamp from the filename

Note: Notebook was performed on workstation

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
import hashlib

data_month_double = '07'

# load data from last notebook
realtime = pd.read_csv('../Data/05_vrs_all_realtime.csv')

realtime.info()

# cut the filename
realtime['filename'] = realtime['filename'].str.split('-', n=1).str[1]

realtime.head()

# remove the last part of the filename
realtime['filename'] = realtime['filename'].str.replace('-UTC_gtfs-realtime-scraper_vrs.csv', '')

realtime.head()

# rename colum to timestamp
realtime.rename(columns={'filename': 'vrs_timestamp'}, inplace=True)

# export
realtime.to_csv('../Data/06_VRS_with_timestamp.csv', index=False)

realtime
