==> Data Preparation/07_Merge_VRS_Trip_Realtime.ipynb <==
# 08 Merge VRS Realtime data with Trips data

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
import hashlib

# Set month
data_month_single = '7'
data_month_double = '07'
data_month_int = 7


# load routes data from notebook 4
routes = pd.read_csv('../Data/04_routes_merged.csv')

## Trips

Prepare trips and merge with routes for later merge with realtime data

# load trips data
trips = pd.DataFrame(pd.read_csv("../Data/01-000000-UTC_gtfs-scraper_trips-vrs.csv"))

# merge the trips and routes DataFrames based on the route_id column
trips = pd.merge(routes, trips, on='route_id')

trips

# trips in the other data have three digits as max
trips['route_id'] = trips['route_id'].astype(str).str[-3:]

# some ids have only two digits
trips['route_id'] = trips['route_id'].astype(str).apply(lambda x: x[-2:] if x[0] == '0' else x)

# read realtime data
realtime = pd.DataFrame(pd.read_csv("../Data/06_VRS_with_timestamp.csv"))


realtime

realtime['route_id'].unique()

realtime.drop_duplicates(inplace=True)

realtime

# filter for trips only being in realtime data
#trips_filtered_on_id = trips[trips['trip_id'].isin(realtime['trip_id'])]

# change route_id format to 
#trips_filtered_on_id['route_id'] = trips_filtered_on_id['route_id'].astype(str).str.slice(3, 6)

#trips_filtered_on_id['route_id'] = trips_filtered_on_id['route_id'].astype(int)

#merged_df = trips_filtered_on_id.merge(realtime, on=['trip_id', 'route_id'])
trips['route_id'] = trips['route_id'].astype(int)


merged_df = trips.merge(realtime, on=['trip_id', 'route_id'])

#export to csv
merged_df.to_csv('../Data/07_vrs_merged_with_trips_NEU.csv', index=False)

merged_df

merged_df['route_id'].unique()
