==> Data Preparation/08_Merge_and_Delay.ipynb <==
# 08: Merge "Stop Times with Buffer zones" with realtime data and compute delays

Here, we will merge the stop times with buffer zones from notebook 4 with the realtime data from notebook 7. Then, we will compute the delay between the actual PT times and the planned ones.

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
import hashlib

# Set month
data_month_single = '7'
data_month_double = '07'
data_month_int = 7

# load PT plan data
plan = pd.read_parquet('../Data/04_merged_stop_times_buffer_zones.parquet', engine='pyarrow')

# transform geometry
point = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(plan["station_point"], crs=4326))
buffer = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(plan["buffer_zone"], crs=4326))

# drop geometry columns in wrong format
plan = plan.drop(columns=["station_point", "buffer_zone"])

# replace geometry columns with correct format
plan['station_point'] = point
plan['buffer_zone'] = buffer

# drop lat and long columns
plan.drop(columns=['lat', 'long'], inplace=True)

# load realtime data
actual = pd.read_csv('../Data/07_vrs_merged_with_trips.csv')

# drop schedule_relationship column, since all values are 0
actual.drop(columns=['schedule_relationship'], inplace=True)

actual.info()

# drop duplicates
actual.drop_duplicates(inplace=True)

actual.info()

actual['route_short_name'].unique()

# merge realtime and plan data
#merged_df = actual.merge(plan, on=['stop_id', 'trip_id'])
merged_df = pd.merge(actual, plan, on=['stop_id', 'trip_id'])

merged_df['route_short_name'].unique()  

# tranform start_date to date
merged_df['start_date'] = pd.to_datetime(merged_df['start_date'], format='%Y%m%d')

# create datetime columns
merged_df['start_datetime'] = pd.to_datetime(merged_df['start_date']) + pd.to_timedelta(merged_df['arrival_time'])
merged_df['end_datetime'] = pd.to_datetime(merged_df['start_date']) + pd.to_timedelta(merged_df['departure_time'])

temporary_gdf = gpd.GeoDataFrame(merged_df, geometry='station_point')

# export to parquet
#temporary_gdf.to_parquet('/Volumes/T7/Master/Processed Data/'+data_month_double+'/08_temporary.parquet')

## Compute Delays

# rename columns
merged_df.rename(columns={'start_datetime': 'scheduled_arrival_time', 'end_datetime': 'scheduled_departure_time'}, inplace=True)
merged_df.rename(columns={'stop_arrival_time': 'actual_arrival_time', 'stop_departure_time': 'actual_departure_time'}, inplace=True)

# drop columns
merged_df.drop(columns=['arrival_time', 'departure_time', 'start_date'], inplace=True)

# replace 0 values with None for datetime conversion
merged_df['actual_arrival_time'] = merged_df['actual_arrival_time'].replace('0', None)
merged_df['actual_departure_time'] = merged_df['actual_departure_time'].replace('0', None)

# convert into datetime
merged_df['actual_arrival_time'] = pd.to_datetime(merged_df['actual_arrival_time'])
merged_df['actual_departure_time'] = pd.to_datetime(merged_df['actual_departure_time'])

# compute delays
merged_df['arrival_delay'] = merged_df['actual_arrival_time'] - merged_df['scheduled_arrival_time']
merged_df['departure_delay'] = merged_df['actual_departure_time'] - merged_df['scheduled_departure_time']

# there are cases, where the actual arrival time is before the scheduled arrival time, meaning that the PT vehicle arrived too early. Set these cases to 0
merged_df['arrival_delay'] = merged_df['arrival_delay'].apply(lambda x: pd.Timedelta(0) if x < pd.Timedelta(0) else x)

# there are cases, where the actual departure time is before the scheduled departure time, meaning that the PT vehicle departed too early. This is not good, but not of interested since it is no delay. Set these cases to 0.
merged_df['departure_delay'] = merged_df['departure_delay'].apply(lambda x: pd.Timedelta(0) if x < pd.Timedelta(0) else x)

# create a GeoDataFrame
gdf = gpd.GeoDataFrame(merged_df, geometry='station_point')

# convert route_short_name to string
gdf['route_short_name'] = gdf['route_short_name'].astype(str)

gdf.to_csv('../Data/08_realtime_buffer_delay.csv', index=False)

# export to parquet
gdf.to_parquet('../Data/08_first_merge_NEU.parquet')
