==> Data Preparation/10_Latest_VRS.ipynb <==
# 10 Keep latest VRS trip data

The vrs data contain different status updates, when the PT vehicle will arrive. Because of that, there are duplicates of information, which need to be removed. 

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
import hashlib

data_month_double = '07'

#data = pd.read_parquet('/Volumes/T7/Master/Processed Data/'+data_month_double+'/10_second_merge.parquet', engine='pyarrow')
#data = pd.read_csv('../Data/10_second_merge.csv')
data = pd.read_csv('../Data/09_second_merge.csv')

data.info()

## Remove

# sort data
sorted_data = data.sort_values(by=['trip_id', 'scheduled_arrival_time'])

sorted_data

# get first trip id
sorted_data.iloc[0]['trip_id']

# get data with first trip id
first_id = data[data['trip_id'] == sorted_data.iloc[0]['trip_id']]

# drop unneccessary columns for better readability
first_id.drop(columns=[ 'route_id', 'route_short_name', 'route_type_name', 'agency_name', 'transfer_stop'], inplace=True)

first_id

# look for duplicates in scheduled arrival time
#duplicates = df[df['scheduled_arrival_time'].duplicated()]

# drop columns for better readability
#duplicates.drop(columns=['agency_id', 'route_type', 'shape_id', 'date'], inplace=True)
first_id.drop(columns=['agency_id', 'route_type', 'shape_id', 'date'], inplace=True)

first_id

# sort 
first_id_sort_by_time = first_id.sort_values(by='scheduled_arrival_time')


# look at specfici stop id
stop_id = first_id_sort_by_time[first_id_sort_by_time['stop_id'] == 7131]

stop_id

sorted_stop_id = stop_id.sort_values(by=['scheduled_arrival_time', 'vrs_timestamp'])

sorted_stop_id

# only look at a specific time
spec_time= sorted_stop_id[sorted_stop_id['scheduled_arrival_time'] == '2023-07-03 05:04:00']

spec_time

spec_time.sort_values(by='vrs_timestamp', ascending=False)

spectime = spec_time.drop_duplicates(subset=['trip_id', 'stop_id', 'scheduled_arrival_time'], keep='last')

spectime

As we can see, there are duplicated columns, that only differ in vrs_timestamp. We need to keep the row with the biggest value in vrs_timestamp, since it represents the latest information about the trip. 

Now: need to develop algorithm to check for duplicates in the specified columns and checking for the biggest value.

Idea: only look at columes used above. Sort the main dataframe by trip_id, scheduled_arrival_time, stop_id, and vrs_timestamp.
Run through the rows and check if charactersics are same, except for vrs_timestamp. For it, check if the true statements in the next line a given and if the timestamp is bigger. If true, keep going, if false, we found the biggest stamp.

Either delete the "current" row if a bigger one was found or search until found and only keep it. 

## Actual 

data_sorted = data.sort_values(by=['trip_id', 'stop_id', 'scheduled_arrival_time', 'vrs_timestamp'])

data_sorted


#data_clean = data_sorted.sort_values('vrs_timestamp', ascending=False)
data_clean = data_sorted.drop_duplicates(subset=['trip_id', 'stop_id', 'scheduled_arrival_time'], keep='last')

data_clean

from shapely.wkt import loads

data_clean['station_point'] = data_clean['station_point'].apply(shapely.wkt.loads)

gdf = gpd.GeoDataFrame(data_clean, geometry='station_point')

#gdf.to_parquet('/Volumes/T7/Master/Processed Data/'+data_month_double+'/11_data_clean.parquet', engine='pyarrow')

#gdf.to_csv('../Data/11_data_clean.csv')
gdf.to_csv('../Data/Data/10_VRS_clean.csv')
