==> Data Preparation/13_Tier_Part3.ipynb <==
# 13: Merge Tier data with VRS data

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
from shapely import Point, wkb, wkt

data_month_single = '7'
data_month_double = '07'

## Load data

# load tier trip data
tier_trips = pd.DataFrame(pd.read_parquet("../Data/12_tier_part2.parquet", engine="pyarrow"))
tier_trips

# transform geometry
point_start = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(tier_trips["start_location"], crs=4326))
point_end = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(tier_trips["end_location"], crs=4326))

# drop geometry columns in wrong format
tier_trips = tier_trips.drop(columns=["start_location", "end_location"])

# replace geometry columns with correct format
tier_trips['start_location'] = point_start
tier_trips['end_location'] = point_end

tier_trips

# transform start_location to geometry
tier_trips = gpd.GeoDataFrame(tier_trips, geometry='start_location')

# transform end_location to geometry
tier_trips = gpd.GeoDataFrame(tier_trips, geometry='end_location')

# set start_location back to GDF geometry
tier_trips = gpd.GeoDataFrame(tier_trips, geometry='start_location')

tier_trips.info()

tier_trips

#vrs_data = pd.DataFrame(pd.read_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/11_data_clean.parquet", engine="pyarrow"))
vrs_data = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/10_VRS_clean.csv"))

# transform geometry
vrs_data['station_point'] = vrs_data['station_point'].apply(wkt.loads)
vrs_data['buffer_zone'] = shapely.wkt.loads(vrs_data['buffer_zone'])
vrs_data = gpd.GeoDataFrame(vrs_data, geometry='buffer_zone')
vrs_data = gpd.GeoDataFrame(vrs_data, geometry='station_point')

# transform geometry
#point = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(vrs_data["station_point"], crs=4326))
#buffer = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(vrs_data["buffer_zone"], crs=4326))

# drop geometry columns in wrong format
#vrs_data= vrs_data.drop(columns=["station_point", "buffer_zone"])

# replace geometry columns with correct format
#vrs_data['station_point'] = point
#vrs_data['buffer_zone'] = buffer

# TO DO: Match Start times

## Match Tier Data with VRS Data

### A: Count entries
A1: Amount of trips starting within timeframe of x minutes
A2: Amount of those trips, that ended at another stations


#tier_trips_start_time
#scheduled_departure_time

vrs = vrs_data.copy()

# convert to datetime
vrs['actual_arrival_time'] =  pd.to_datetime(vrs['actual_arrival_time'])
vrs['actual_departure_time'] =  pd.to_datetime(vrs['actual_departure_time'])
vrs['scheduled_arrival_time'] =  pd.to_datetime(vrs['scheduled_arrival_time'])
vrs['scheduled_departure_time'] =  pd.to_datetime(vrs['scheduled_departure_time'])

# keep only tier data where a trip staerted at a station
tier_trips_not_nan = tier_trips[tier_trips['start_stop_id'].notna()]

# Slice the dataframe into three equal parts for easier processing
num_rows = len(vrs)
part_size = num_rows // 3

# First part
part1 = vrs.iloc[:part_size]

# Second part
part2 = vrs.iloc[part_size:2*part_size]

# Third part
part3 = vrs.iloc[2*part_size:]

# Create a new column in vrs_data to store the count
part1['tier_trips_count'] = 0
part1['tier_trips_id'] = None
part1['tier_trips_end_at_station_count'] = 0
part1['tier_trips_end_at_station_id'] = None

# Iterate over each row in vrs_data
for index, row in part1.iterrows():
    # Get the stop_id and scheduled_arrival_time
    stop_id = row['stop_id']
    scheduled_departure_time = row['scheduled_departure_time']
    
    # Calculate the time span
    time_span = pd.Timedelta(minutes=5)
    
    # Filter tier_trips based on stop_id and time span
    filtered_trips = tier_trips_not_nan[(tier_trips_not_nan['start_stop_id'] == stop_id) & 
                                (tier_trips_not_nan['tier_trips_start_time'] >= scheduled_departure_time) & 
                                (tier_trips_not_nan['tier_trips_start_time'] <= scheduled_departure_time + time_span)]
    
    filtered_ids = filtered_trips['tier_trips_id'].values
    
    # Count the number of entries
    count = len(filtered_ids)
    
    # Update the count in vrs_data
    part1.at[index, 'tier_trips_count'] = count
    part1.at[index, 'tier_trips_id'] = filtered_ids # list of tier_trips ids

    # Trips that ended at station
    end_trips = filtered_trips[filtered_trips['end_stop_id'].notna()]
    end_trips_ids = end_trips['tier_trips_id'].values
    end_trips_count = len(end_trips_ids)

    part1.at[index, 'tier_trips_end_at_station_count'] = end_trips_count
    part1.at[index, 'tier_trips_end_at_station_id'] = end_trips_ids

# create a GeoDataFrame
gdf1 = gpd.GeoDataFrame(part1, geometry='station_point')

#
#gdf1 = gdf1[gdf1['tier_trips_count'] != 0]

# convert to string
gdf1['tier_trips_id'] = gdf1['tier_trips_id'].astype(str)

# remove brackets
gdf1['tier_trips_id'] = gdf1['tier_trips_id'].str.replace('[', '').str.replace(']', '')

# convert to string
gdf1['tier_trips_end_at_station_id'] = gdf1['tier_trips_end_at_station_id'].astype(str)
gdf1['route_short_name'] = gdf1['route_short_name'].astype(str)

# export to parquet
gdf1.to_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/13_VRS_Tier_Merge_1_NEU_NEU_TEST.parquet", engine="pyarrow")

# Create a new column in vrs_data to store the count
part2['tier_trips_count'] = 0
part2['tier_trips_id'] = None
part2['tier_trips_end_at_station_count'] = 0
part2['tier_trips_end_at_station_id'] = None

# Iterate over each row in vrs_data
for index, row in part2.iterrows():
    # Get the stop_id and scheduled_arrival_time
    stop_id = row['stop_id']
    scheduled_arrival_time = row['scheduled_departure_time']
    
    # Calculate the time span
    time_span = pd.Timedelta(minutes=5)
    
    # Filter tier_trips based on stop_id and time span
    filtered_trips = tier_trips_not_nan[(tier_trips_not_nan['start_stop_id'] == stop_id) & 
                                (tier_trips_not_nan['tier_trips_start_time'] >= scheduled_arrival_time) & 
                                (tier_trips_not_nan['tier_trips_start_time'] <= scheduled_arrival_time + time_span)]
    
    filtered_ids = filtered_trips['tier_trips_id'].values
    
    # Count the number of entries
    count = len(filtered_ids)
    
    # Update the count in vrs_data
    part2.at[index, 'tier_trips_count'] = count
    part2.at[index, 'tier_trips_id'] = filtered_ids # list of tier_trips ids

    # Trips that ended at station
    end_trips = filtered_trips[filtered_trips['end_stop_id'].notna()]
    end_trips_ids = end_trips['tier_trips_id'].values
    end_trips_count = len(end_trips_ids)

    part2.at[index, 'tier_trips_end_at_station_count'] = end_trips_count
    part2.at[index, 'tier_trips_end_at_station_id'] = end_trips_ids

gdf2 = gpd.GeoDataFrame(part2, geometry='station_point')

#gdf2 = gdf2[gdf2['tier_trips_count'] != 0]

gdf2['tier_trips_id'] = gdf2['tier_trips_id'].astype(str)

gdf2['tier_trips_id'] = gdf2['tier_trips_id'].str.replace('[', '').str.replace(']', '')

gdf2['tier_trips_end_at_station_id'] = gdf2['tier_trips_end_at_station_id'].astype(str)
gdf2['route_short_name'] = gdf2['route_short_name'].astype(str)
gdf2.to_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/13_VRS_Tier_Merge_2_NEU_NEU_TEST.parquet", engine="pyarrow")

# Create a new column in vrs_data to store the count
part3['tier_trips_count'] = 0
part3['tier_trips_id'] = None
part3['tier_trips_end_at_station_count'] = 0
part3['tier_trips_end_at_station_id'] = None

# Iterate over each row in vrs_data
for index, row in part3.iterrows():
    # Get the stop_id and scheduled_arrival_time
    stop_id = row['stop_id']
    scheduled_departure_time = row['scheduled_departure_time']
    
    # Calculate the time span
    time_span = pd.Timedelta(minutes=5)
    
    # Filter tier_trips based on stop_id and time span
    filtered_trips = tier_trips_not_nan[(tier_trips_not_nan['start_stop_id'] == stop_id) & 
                                (tier_trips_not_nan['tier_trips_start_time'] >= scheduled_departure_time) & 
                                (tier_trips_not_nan['tier_trips_start_time'] <= scheduled_departure_time + time_span)]
    
    filtered_ids = filtered_trips['tier_trips_id'].values
    
    # Count the number of entries
    count = len(filtered_ids)
    
    # Update the count in vrs_data
    part3.at[index, 'tier_trips_count'] = count
    part3.at[index, 'tier_trips_id'] = filtered_ids # list of tier_trips ids

    # Trips that ended at station
    end_trips = filtered_trips[filtered_trips['end_stop_id'].notna()]
    end_trips_ids = end_trips['tier_trips_id'].values
    end_trips_count = len(end_trips_ids)

    part3.at[index, 'tier_trips_end_at_station_count'] = end_trips_count
    part3.at[index, 'tier_trips_end_at_station_id'] = end_trips_ids

gdf3 = gpd.GeoDataFrame(part3, geometry='station_point')

#gdf3 = gdf3[gdf3['tier_trips_count'] != 0]

gdf3['tier_trips_id'] = gdf3['tier_trips_id'].astype(str)

gdf3['tier_trips_id'] = gdf3['tier_trips_id'].str.replace('[', '').str.replace(']', '')

gdf3['tier_trips_end_at_station_id'] = gdf3['tier_trips_end_at_station_id'].astype(str)
gdf3['route_short_name'] = gdf3['route_short_name'].astype(str)
gdf3.to_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/13_VRS_Tier_Merge_3_NEU_NEU_TEST.parquet", engine="pyarrow")

# combine the three GeoDataFrames
combined = pd.concat([gdf1, gdf2, gdf3])

# export to parquet
combined.to_parquet("../Data/13_VRS_Tier_Merge_Complete.parquet", engine="pyarrow")

combined

tier_trips.columns

entries = tier_trips[(tier_trips['start_stop_id'].notna()) & (tier_trips['end_stop_id'].notna()) & (tier_trips['start_stop_id'] != tier_trips['end_stop_id'])]
entries

# REWORK WITH TIER DATASET - AKTUELL EIN FEHLER - ALLE STARTEN UND ENDEN AN STATION

entries = combined[(combined['tier_trips_count'] > 0) & (combined['tier_trips_end_at_station_count'] > 0) & (combined['tier_trips_count'] != combined['tier_trips_end_at_station_count'])]
entries
