==> Data Preparation/14_Next_Part1.ipynb <==
# 13: Nextbike part 1

*Note: The Nextbike Notebooks are identical to the Tier notebooks*

In this notebook, we will prepare the Nextbike dataset. This includes
- checking for duplicates
- checking for anomalies in rides e.g. same start and end coordinates, checking same start and end times
- calcualte trip durations
- calculate trip distances
- create "vehicle type" colum

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium

data_month_single = '7'
data_month_double = '07'

## load data

# load tier trip data
nextbike = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/Raw Data/nextbike_trips_bonn.csv"))
nextbike

# check data types
nextbike.info()

gdf = gpd.GeoDataFrame(nextbike)

# get lat an long values from string
def extract_coordinates(point_str):
    point_str = point_str.replace("POINT (", "").replace(")", "")  # Remove "POINT (" and ")"
    x_start, y_start = point_str.split()  # Split the string by whitespace
    return float(x_start), float(y_start)

# add x and y columns to gdf
gdf[['x_start', 'y_start']] = gdf['nextbike_trips_start_geometry'].apply(lambda point_str: pd.Series(extract_coordinates(point_str)))


# set geometry
gdf = gpd.GeoDataFrame(gdf, geometry=gpd.points_from_xy(gdf.y_start, gdf.x_start), crs="EPSG:4326")

# drop columns
gdf.drop(columns=['nextbike_trips_start_geometry', 'x_start', 'y_start'], inplace=True)

# rename
gdf = gdf.rename(columns={'geometry': 'start_location'})

gdf[['x_end', 'y_end']] = gdf['nextbike_trips_end_geometry'].apply(lambda point_str: pd.Series(extract_coordinates(point_str)))

gdf = gpd.GeoDataFrame(gdf, geometry=gpd.points_from_xy(gdf.y_end, gdf.x_end), crs="EPSG:4326")

gdf = gdf.rename(columns={'geometry': 'end_location'})

gdf.drop(columns=['x_end', 'y_end'], inplace=True)

gdf.drop(columns=['nextbike_trips_end_geometry'], inplace=True)

gdf

# convert to datetime
gdf['nextbike_trips_start_time'] = pd.to_datetime(gdf['nextbike_trips_start_time'])
gdf['nextbike_trips_end_time'] = pd.to_datetime(gdf['nextbike_trips_end_time'])

#check if all month are included in dataset
months = gdf['nextbike_trips_start_time'].dt.month.unique()
print(months)

# set new column with vehicle time - important if bikes are added later on 
gdf['vehicle type'] = 'bike'

# sort dataframe by start times
gdf_sorted= gdf.sort_values('nextbike_trips_start_time')
gdf_sorted

# Check for duplicates in the 'tier' DataFrame
duplicates_exist = gdf.duplicated().any()

if duplicates_exist:
    print("Duplicates exist in the DataFrame")
else:
    print("No duplicates found in the DataFrame")

# check if trips are starting and ending at the same location
gdf_same_location = gdf[gdf['start_location'] == gdf['end_location']]
gdf_same_location

# only keep entries where start_location and end_location are not identical
gdf = gdf[gdf['start_location'] != gdf['end_location']]
gdf

#check if there are trips with the same start and end time
gdf_same_time = gdf[gdf['nextbike_trips_start_time'] == gdf['nextbike_trips_end_time']]
gdf_same_time

# only keep entries where start and end tieme are not identical
gdf = gdf[gdf['nextbike_trips_start_time'] != gdf['nextbike_trips_end_time']]
gdf

# calculate trip duration
gdf['trip_duration'] = gdf['nextbike_trips_end_time'] - gdf['nextbike_trips_start_time']

gdf

#look at trip durations
trip_durations = gdf['nextbike_trips_end_time'] - gdf['nextbike_trips_start_time']
unique_durations = trip_durations.unique()
print(unique_durations)

We can see, that the shortest trips are 5 minutes long, while the longest trips are over 20 days long

# only keep trips with rental duration of less than 2 hours
gdf = gdf[gdf['trip_duration'] <= pd.Timedelta(hours=2)]


# only keep rentals longer than 1 minute
gdf = gdf[gdf['trip_duration'] >= pd.Timedelta(minutes=1)]

gdf

#calculate driven distance
# code from ChatGPT
import math

def haversine_distance(lat1, lon1, lat2, lon2):
    # Convert latitude and longitude to radians
    lat1_rad = math.radians(lat1)
    lon1_rad = math.radians(lon1)
    lat2_rad = math.radians(lat2)
    lon2_rad = math.radians(lon2)

    # Haversine formula
    dlon = lon2_rad - lon1_rad
    dlat = lat2_rad - lat1_rad
    a = math.sin(dlat/2)**2 + math.cos(lat1_rad) * math.cos(lat2_rad) * math.sin(dlon/2)**2
    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1-a))
    distance = 6371 * c  # Earth's radius in kilometers
    return distance

# Calculate driven distance for each trip
gdf['trip_distance_kilometers'] = gdf.apply(lambda row: haversine_distance(row['start_location'].y, row['start_location'].x, row['end_location'].y, row['end_location'].x), axis=1)

gdf

# Calculate driven distance in meters
gdf['trip_distance_meters'] = gdf['trip_distance_kilometers']*1000

# delete all trips with a distance of more than 15 km according to Reck et al. (2021)
gdf = gdf[gdf['trip_distance_meters'] <= 15000]

# only keep trips with distance of more than 100 meters
gdf = gdf[gdf['trip_distance_meters'] >= 100]
gdf

gdf.info()

gdf.isnull().sum()


# Only 2023 data
gdf_2023 = gdf[gdf['nextbike_trips_start_time'] <= '2023-12-31']
gdf_2023 = gdf_2023[gdf_2023['nextbike_trips_end_time'] <= '2023-12-31']
gdf_2023

gdf_2023.info()

gdf_2023.to_parquet('/Volumes/T7/Master/Processed Data/'+data_month_double+'/14_nextbike_part1.parquet') 

gdf_2023
