==> Data Preparation/15_Next_Part2.ipynb <==
# 15: Nextbike data part 2 - get which Nextbike trips started/ended at stations

In this notebook, we will look, which Nextbike trips started within one of our stop buffer zones. If a trip started (or ended or both) in such a buffer zone, the information will be added to the Nextbike dataset
- start station 
    - id
    - name
    - buffer zone
- end station
    - id
    - name
    - buffer zone

Note: Originally on local machine performed

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
from shapely import Point, wkb, wkt

data_month_single = '7'
data_month_double = '07'

# load nextbike trip data
nextbike = pd.DataFrame(pd.read_parquet("../Data/14_nextbike_part1.parquet", engine="pyarrow"))
nextbike

# transform geometry
point_start = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(nextbike["start_location"], crs=4326))
point_end = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(nextbike["end_location"], crs=4326))

# drop geometry columns in wrong format
nextbike = nextbike.drop(columns=["start_location", "end_location"])

# replace geometry columns with correct format
nextbike['start_location'] = point_start
nextbike['end_location'] = point_end

nextbike

nextbike['start_location'] = nextbike['start_location'].apply(lambda point: Point(point.y, point.x))
nextbike['end_location'] = nextbike['end_location'].apply(lambda point: Point(point.y, point.x))

nextbike

nextbike.info()

nextbike = gpd.GeoDataFrame(nextbike, geometry='start_location')

nextbike = gpd.GeoDataFrame(nextbike, geometry='end_location')

nextbike = gpd.GeoDataFrame(nextbike, geometry='start_location')

nextbike.info()

nextbike

#vrs_data = pd.DataFrame(pd.read_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/11_data_clean.parquet", engine="pyarrow"))

vrs_data = pd.DataFrame(pd.read_parquet("../Data/10_VRS_data_clean.parquet", engine="pyarrow"))
#vrs_data = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/19_Merged_with_Weather_Holiday.csv"))
#vrs_data = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/10_VRS_data_clean.csv"))

# transform geometry
vrs_data['station_point'] = vrs_data['station_point'].apply(wkt.loads)

vrs_data['buffer_zone'] = shapely.wkt.loads(vrs_data['buffer_zone'])

vrs_data = gpd.GeoDataFrame(vrs_data, geometry='station_point')

# transform geometry
#point = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(vrs_data["station_point"], crs=4326))
#buffer = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(vrs_data["buffer_zone"], crs=4326))

# drop geometry columns in wrong format
#vrs_data= vrs_data.drop(columns=["station_point", "buffer_zone"])

# replace geometry columns with correct format
#vrs_data['station_point'] = point
#vrs_data['buffer_zone'] = buffer

vrs_data.info()

## Check if Tier trip start and/or ends within a buffer zone

unique_stops = vrs_data[['stop_id', 'stop_name', 'buffer_zone']].drop_duplicates()

# check if trip starts in a buffer zone

# Create empty columns in tier_data
nextbike['start_stop_id'] = np.nan
nextbike['start_stop_name'] = np.nan
nextbike['start_buffer_zone'] = np.nan

# Iterate over each row in tier_data
for index, row in nextbike.iterrows():
    # Get the coordinates of tier_trips_end
    tier_trips_end_coords = row['start_location']	
    
    # Check if tier_trips_end coordinates lie within a buffer_zone in unique_stops
    for stop_index, stop_row in unique_stops.iterrows():
        buffer_zone = stop_row['buffer_zone']
        if buffer_zone.contains(tier_trips_end_coords):
            # If true, add the stop_id, stop_name, buffer_zone to the row in tier_data
            nextbike.at[index, 'start_stop_id'] = stop_row['stop_id']
            nextbike.at[index, 'start_stop_name'] = stop_row['stop_name']
            nextbike.at[index, 'start_buffer_zone'] = buffer_zone
            break

# check if trip ends in a buffer zone

# Create empty columns in tier_data
nextbike['end_stop_id'] = np.nan
nextbike['end_stop_name'] = np.nan
nextbike['end_buffer_zone'] = np.nan

# Iterate over each row in tier_data
for index, row in nextbike.iterrows():
    # Get the coordinates of tier_trips_end
    tier_trips_end_coords = row['end_location']	
    
    # Check if tier_trips_end coordinates lie within a buffer_zone in unique_stops
    for stop_index, stop_row in unique_stops.iterrows():
        buffer_zone = stop_row['buffer_zone']
        if buffer_zone.contains(tier_trips_end_coords):
            # If true, add the stop_id, stop_name, buffer_zone to the row in tier_data
            nextbike.at[index, 'end_stop_id'] = stop_row['stop_id']
            nextbike.at[index, 'end_stop_name'] = stop_row['stop_name']
            nextbike.at[index, 'end_buffer_zone'] = buffer_zone
            break

nextbike

nextbike.info()

gdf = gpd.GeoDataFrame(nextbike, geometry='start_buffer_zone')

gdf = gpd.GeoDataFrame(nextbike, geometry='end_buffer_zone')

gdf = gpd.GeoDataFrame(nextbike, geometry='start_location')

gdf.info()


def parse_wkb(x):
    try:
        return wkb.loads(str(x), hex=True) if x else None
    except:
        return None

gdf['start_buffer_zone'] = gdf['start_buffer_zone'].apply(parse_wkb)


gdf

#nextbike['start_buffer_zone'] = nextbike['start_buffer_zone'].apply(lambda x: x.wkb if isinstance(x, gpd.GeoSeries) else None)



#nextbike['end_buffer_zone'] = nextbike['end_buffer_zone'].apply(lambda x: x.wkb if isinstance(x, gpd.GeoSeries) else None)

gdf.info()

nextbike.info()

nextbike['start_buffer_zone'] = str(nextbike['start_buffer_zone'])
nextbike['end_buffer_zone'] = str(nextbike['end_buffer_zone'])

nextbike.to_parquet("../Data/15_nextbike_part2.parquet", engine="pyarrow")

nextbike_not_nan = nextbike[nextbike['start_stop_id'].notna()]
nextbike_not_nan
