==> Data Preparation/16_Next_Part3.ipynb <==
# 16: Merge Nextbike data with VRS data

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
from shapely import Point, wkb, wkt

data_month_single = '7'
data_month_double = '07'

## load data

# load tier trip data
nextbike = pd.DataFrame(pd.read_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/15_nextbike_with_buffer_zones_NEU.parquet", engine="pyarrow"))
nextbike

# transform geometry
point_start = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(nextbike["start_location"], crs=4326))
point_end = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(nextbike["end_location"], crs=4326))

# drop geometry columns in wrong format
nextbike = nextbike.drop(columns=["start_location", "end_location"])

# replace geometry columns with correct format
nextbike['start_location'] = point_start
nextbike['end_location'] = point_end

nextbike

nextbike = gpd.GeoDataFrame(nextbike, geometry='start_location')

nextbike = gpd.GeoDataFrame(nextbike, geometry='end_location')

nextbike = gpd.GeoDataFrame(nextbike, geometry='start_location')

nextbike.info()

nextbike

#vrs_data = pd.DataFrame(pd.read_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/10_VRS_data_clean.parquet", engine="pyarrow"))
vrs_data = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/11_data_clean.csv"))
#vrs_data = pd.DataFrame(pd.read_csv("/Volumes/T7/Master/10_VRS_data_clean.csv"))

# transform geometry
vrs_data['station_point'] = vrs_data['station_point'].apply(wkt.loads)
vrs_data['buffer_zone'] = shapely.wkt.loads(vrs_data['buffer_zone'])
vrs_data = gpd.GeoDataFrame(vrs_data, geometry='buffer_zone')
vrs_data = gpd.GeoDataFrame(vrs_data, geometry='station_point')

# transform geometry
#point = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(vrs_data["station_point"], crs=4326))
#buffer = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(vrs_data["buffer_zone"], crs=4326))

# drop geometry columns in wrong format
#vrs_data= vrs_data.drop(columns=["station_point", "buffer_zone"])

# replace geometry columns with correct format
#vrs_data['station_point'] = point
#vrs_data['buffer_zone'] = buffer

nextbike_trips_not_nan = nextbike[nextbike['start_stop_id'].notna()]

nextbike_trips_not_nan

vrs = vrs_data.copy()

vrs['actual_arrival_time'] =  pd.to_datetime(vrs['actual_arrival_time'])
vrs['actual_departure_time'] =  pd.to_datetime(vrs['actual_departure_time'])
vrs['scheduled_arrival_time'] =  pd.to_datetime(vrs['scheduled_arrival_time'])
vrs['scheduled_departure_time'] =  pd.to_datetime(vrs['scheduled_departure_time'])

# Assuming you have a dataframe called 'df'

# Slice the dataframe into three equal parts
num_rows = len(vrs)
part_size = num_rows // 3

# First part
part1 = vrs.iloc[:part_size]

# Second part
part2 = vrs.iloc[part_size:2*part_size]

# Third part
part3 = vrs.iloc[2*part_size:]

# Create a new column in vrs_data to store the count
part1['nextbike_trips_count'] = 0
part1['nextbike_trips_id'] = None
part1['nextbike_trips_end_at_station_count'] = 0
part1['nextbike_trips_end_at_station_id'] = None

# Iterate over each row in vrs_data
for index, row in part1.iterrows():
    # Get the stop_id and scheduled_arrival_time
    stop_id = row['stop_id']
    scheduled_arrival_time = row['scheduled_departure_time']
    
    # Calculate the time span
    time_span = pd.Timedelta(minutes=5)
    
    # Filter tier_trips based on stop_id and time span
    filtered_trips = nextbike_trips_not_nan[(nextbike_trips_not_nan['start_stop_id'] == stop_id) & 
                                (nextbike_trips_not_nan['nextbike_trips_start_time'] >= scheduled_arrival_time) & 
                                (nextbike_trips_not_nan['nextbike_trips_start_time'] <= scheduled_arrival_time + time_span)]
    
    filtered_ids = filtered_trips['nextbike_trips_id'].values
    
    # Count the number of entries
    count = len(filtered_ids)
    
    # Update the count in vrs_data
    part1.at[index, 'nextbike_trips_count'] = count
    part1.at[index, 'nextbike_trips_id'] = filtered_ids # list of tier_trips ids

    # Trips that ended at station
    end_trips = filtered_trips[filtered_trips['end_stop_id'].notna()]
    end_trips_ids = end_trips['nextbike_trips_id'].values
    end_trips_count = len(end_trips_ids)

    part1.at[index, 'nextbike_trips_end_at_station_count'] = end_trips_count
    part1.at[index, 'nextbike_trips_end_at_station_id'] = end_trips_ids

gdf1 = gpd.GeoDataFrame(part1, geometry='station_point')

#gdf1 = gdf1[gdf1['nextbike_trips_count'] != 0]

gdf1['nextbike_trips_id'] = gdf1['nextbike_trips_id'].astype(str)

gdf1['nextbike_trips_id'] = gdf1['nextbike_trips_id'].str.replace('[', '').str.replace(']', '')

gdf1['nextbike_trips_end_at_station_id'] = gdf1['nextbike_trips_end_at_station_id'].astype(str)
gdf1['route_short_name'] = gdf1['route_short_name'].astype(str)
gdf1.to_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/16_VRS_Tier_Next_Merge_1_NEU_NEU_TEST.parquet", engine="pyarrow")

# Create a new column in vrs_data to store the count
part2['nextbike_trips_count'] = 0
part2['nextbike_trips_id'] = None
part2['nextbike_trips_end_at_station_count'] = 0
part2['nextbike_trips_end_at_station_id'] = None

# Iterate over each row in vrs_data
for index, row in part2.iterrows():
    # Get the stop_id and scheduled_arrival_time
    stop_id = row['stop_id']
    scheduled_arrival_time = row['scheduled_departure_time']
    
    # Calculate the time span
    time_span = pd.Timedelta(minutes=5)
    
    # Filter tier_trips based on stop_id and time span
    filtered_trips = nextbike_trips_not_nan[(nextbike_trips_not_nan['start_stop_id'] == stop_id) & 
                                (nextbike_trips_not_nan['nextbike_trips_start_time'] >= scheduled_arrival_time) & 
                                (nextbike_trips_not_nan['nextbike_trips_start_time'] <= scheduled_arrival_time + time_span)]
    
    filtered_ids = filtered_trips['nextbike_trips_id'].values
    
    # Count the number of entries
    count = len(filtered_ids)
    
    # Update the count in vrs_data
    part2.at[index, 'nextbike_trips_count'] = count
    part2.at[index, 'nextbike_trips_id'] = filtered_ids # list of tier_trips ids

     # Trips that ended at station
    end_trips = filtered_trips[filtered_trips['end_stop_id'].notna()]
    end_trips_ids = end_trips['nextbike_trips_id'].values
    end_trips_count = len(end_trips_ids)

    part2.at[index, 'nextbike_trips_end_at_station_count'] = end_trips_count
    part2.at[index, 'nextbike_trips_end_at_station_id'] = end_trips_ids

gdf2 = gpd.GeoDataFrame(part2, geometry='station_point')

#gdf2 = gdf2[gdf2['nextbike_trips_count'] != 0]

gdf2['nextbike_trips_id'] = gdf2['nextbike_trips_id'].astype(str)

gdf2['nextbike_trips_id'] = gdf2['nextbike_trips_id'].str.replace('[', '').str.replace(']', '')

gdf2

gdf2['nextbike_trips_end_at_station_id'] = gdf2['nextbike_trips_end_at_station_id'].astype(str)
gdf2['route_short_name'] = gdf2['route_short_name'].astype(str)
gdf2.to_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/16_VRS_Tier_Next_Merge_2_NEU_NEU_TEST.parquet", engine="pyarrow")

# Create a new column in vrs_data to store the count
part3['nextbike_trips_count'] = 0
part3['nextbike_trips_id'] = None
part3['nextbike_trips_end_at_station_count'] = 0
part3['nextbike_trips_end_at_station_id'] = None

# Iterate over each row in vrs_data
for index, row in part3.iterrows():
    # Get the stop_id and scheduled_arrival_time
    stop_id = row['stop_id']
    scheduled_arrival_time = row['scheduled_departure_time']
    
    # Calculate the time span
    time_span = pd.Timedelta(minutes=5)
    
    # Filter tier_trips based on stop_id and time span
    filtered_trips = nextbike_trips_not_nan[(nextbike_trips_not_nan['start_stop_id'] == stop_id) & 
                                (nextbike_trips_not_nan['nextbike_trips_start_time'] >= scheduled_arrival_time) & 
                                (nextbike_trips_not_nan['nextbike_trips_start_time'] <= scheduled_arrival_time + time_span)]
    
    filtered_ids = filtered_trips['nextbike_trips_id'].values
    
    # Count the number of entries
    count = len(filtered_ids)
    
    # Update the count in vrs_data
    part3.at[index, 'nextbike_trips_count'] = count
    part3.at[index, 'nextbike_trips_id'] = filtered_ids # list of tier_trips ids

     # Trips that ended at station
    end_trips = filtered_trips[filtered_trips['end_stop_id'].notna()]
    end_trips_ids = end_trips['nextbike_trips_id'].values
    end_trips_count = len(end_trips_ids)

    part3.at[index, 'nextbike_trips_end_at_station_count'] = end_trips_count
    part3.at[index, 'nextbike_trips_end_at_station_id'] = end_trips_ids

gdf3 = gpd.GeoDataFrame(part3, geometry='station_point')

#gdf3 = gdf3[gdf3['nextbike_trips_count'] != 0]

gdf3['nextbike_trips_id'] = gdf3['nextbike_trips_id'].astype(str)

gdf3['nextbike_trips_id'] = gdf3['nextbike_trips_id'].str.replace('[', '').str.replace(']', '')

gdf3['nextbike_trips_end_at_station_id'] = gdf3['nextbike_trips_end_at_station_id'].astype(str)
gdf3['route_short_name'] = gdf3['route_short_name'].astype(str)
gdf3.to_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/16_VRS_Tier_Next_Merge_3_NEU_NEU_TEST.parquet", engine="pyarrow")

combined = pd.concat([gdf1, gdf2, gdf3])

combined['route_short_name'].unique()

combined['route_short_name'] = combined['route_short_name'].astype(str)
combined['nextbike_trips_end_at_station_id'] = combined['nextbike_trips_end_at_station_id'].astype(str)
combined.to_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/16_VRS_Next_Merge_Complete_NEU_NEU_TEST.parquet", engine="pyarrow")
