==> Data Preparation/17_Merge_Tier_Next.ipynb <==
# 17: Merge Tier and Nextbike data

Here we are going to merge the tier and nextbike datasets (which are already merged with VRS)

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
from shapely import Point

data_month_single = '7'
data_month_double = '07'

## load data

tier = pd.DataFrame(pd.read_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/13_VRS_Tier_Merge_Complete_NEU_NEU_TEST.parquet", engine="pyarrow"))

next = pd.DataFrame(pd.read_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/16_VRS_Next_Merge_Complete_NEU_NEU_TEST.parquet", engine="pyarrow"))

# transform geometry
tier_point = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(tier["station_point"], crs=4326))
tier_buffer = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(tier["buffer_zone"], crs=4326))

# drop geometry columns in wrong format
tier = tier.drop(columns=["station_point", "buffer_zone"])

# replace geometry columns with correct format
tier['station_point'] = tier_point
tier['buffer_zone'] = tier_buffer

# transform geometry
next_point = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(next["station_point"], crs=4326))
next_buffer = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(next["buffer_zone"], crs=4326))

# drop geometry columns in wrong format
next = next.drop(columns=["station_point", "buffer_zone"])

# replace geometry columns with correct format
next['station_point'] = next_point
next['buffer_zone'] = next_buffer

## Merge Tier and Next Dataframes

tier.info()

tier.drop(columns=['Unnamed: 0.2', 'Unnamed: 0.1', 'Unnamed: 0'], inplace=True)

tier

next.info()

next.drop(columns=['Unnamed: 0.2', 'Unnamed: 0.1', 'Unnamed: 0'], inplace=True)

next

merge = pd.merge(tier, next, on=['route_id', 'agency_id', 'route_short_name', 'route_type',
       'route_type_name', 'agency_name', 'service_id', 'trip_id',
       'trip_headsign', 'direction_id', 'shape_id', 'stop_id',
       'actual_arrival_time', 'actual_departure_time', 'vrs_timestamp',
       'stop_sequence', 'stop_headsign', 'stop_name', 'scheduled_arrival_time',
       'scheduled_departure_time', 'arrival_delay', 'departure_delay',
       'service', 'date', 'weekday', 'transfer_stop', 'station_point', 'buffer_zone'], how="outer", suffixes=('_tier', '_next'))

merge

gdf = gpd.GeoDataFrame(merge, geometry='station_point')

gdf.to_parquet("/Volumes/T7/Master/Processed Data/"+data_month_double+"/17_VRS_Tier_Next_Merge_Complete_NEU_NEU_TEST.parquet", engine="pyarrow")

gdf
