==> Data Preparation/18_Weather.ipynb <==
# 18: Weather

# import libraries
import pandas as pd
import numpy as np
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
import hashlib

data_month_double = '07'

## load data

vrs_data = pd.read_parquet('/Volumes/T7/Master/Processed Data/'+data_month_double+'/17_VRS_Tier_Next_Merge_Complete_NEU_NEU_TEST.parquet', engine='pyarrow')

# transform geometry
point = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(vrs_data["station_point"], crs=4326))
buffer = gpd.GeoDataFrame(geometry=gpd.GeoSeries.from_wkb(vrs_data["buffer_zone"], crs=4326))

# drop geometry columns in wrong format
vrs_data = vrs_data.drop(columns=["station_point", "buffer_zone"])

# replace geometry columns with correct format
vrs_data['station_point'] = point
vrs_data['buffer_zone'] = buffer

vrs_data

# load weather data
weather_data = pd.read_csv('/Volumes/T7/Master/Raw Data/weather_bonn.csv')

## cut and merge with main data

# only keep actual weather data
weather_data = weather_data.filter(like='current')

# drop column
weather_data.drop(columns=["current_snow.1h"], inplace=True)

weather_data

weather_data.info()

# convert into datetime
weather_data['current_time'] = pd.to_datetime(weather_data['current_time'], utc = True)


# delete timezone
weather_data['current_time'] = weather_data['current_time'].dt.tz_localize(None)



# add one hour to get back to original time (Berlin time)
weather_data['current_time'] = weather_data['current_time'] + pd.Timedelta(hours=1)


# only 2023 data
weather_data = weather_data[weather_data['current_time'].dt.year == 2023]


# sort weather data
sorted_weather_data = weather_data.sort_values('current_time')


# reset index
sorted_weather_data = sorted_weather_data.reset_index(drop=True)

# drop Null values in vrs data (needed to merge with weather)
#vrs_data.dropna(subset=['actual_departure_time', 'actual_arrival_time'], inplace=True)

# sort by actual departure time (needed for merge)
vrs_data_sorted = vrs_data.sort_values('scheduled_arrival_time')


# merge on nearest time
merged_data = pd.merge_asof(vrs_data_sorted, sorted_weather_data, left_on='scheduled_arrival_time', right_on='current_time')

merged_data

gdf = gpd.GeoDataFrame(merged_data, geometry='station_point')

gdf.to_parquet('/Volumes/T7/Master/Processed Data/'+data_month_double+'/18_Merged_with_Weather_NEU_NEU_TEST.parquet')
