==> Data Preparation/20_Extend_Data.ipynb <==
# 20: Extend Dataset

In this notebook, new and initally not planned features are added to the data

import numpy as np
import pandas as pd
import statsmodels.api as sm
import matplotlib.pyplot as plt
import seaborn as sns
import sklearn
from datetime import datetime
import os
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
import shapely
import hashlib
import holidays
from shapely import Point, wkt
import scipy.stats
import seaborn as sns
import statsmodels.api as sm
import statsmodels.formula.api as smf

data_month_double = '07'

## Load data

# read final data
data = pd.read_csv("../Data/18_Merged_with_Weather.csv")

tier = pd.read_parquet("../Data/12_part2.parquet")
next = pd.read_parquet("../Data/15_part2.parquet")

tier['trip_duration_seconds'] = tier['trip_duration'].dt.total_seconds()
next['trip_duration_seconds'] = next['trip_duration'].dt.total_seconds()

# transform geometry
data['station_point'] = data['station_point'].apply(wkt.loads)

data['buffer_zone'] = shapely.wkt.loads(data['buffer_zone'])

data = gpd.GeoDataFrame(data, geometry='station_point')

data = data.loc[:, ~data.columns.str.contains('Unnamed')]

#data.drop(columns=['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.2'], inplace=True)

data['actual_arrival_time'] = pd.to_datetime(data['actual_arrival_time'])
data['actual_departure_time'] = pd.to_datetime(data['actual_departure_time'])
data['scheduled_arrival_time'] = pd.to_datetime(data['scheduled_arrival_time'])
data['scheduled_departure_time'] = pd.to_datetime(data['scheduled_departure_time'])

data = gpd.GeoDataFrame(data, geometry='buffer_zone')

data['date'] = pd.to_datetime(data['date'])

data['arrival_delay'] = pd.to_timedelta(data['arrival_delay'])
data['departure_delay'] = pd.to_timedelta(data['departure_delay'])

## Add features

data['arrival_delay_float'] = data['arrival_delay'].astype(str)
data['arrival_delay_float'] = data['arrival_delay_float'].str.split(' ', n=2).str[-1]
data['arrival_delay_float'] = data['arrival_delay_float'].astype(str)
data['arrival_delay_float'] = pd.to_timedelta(data['arrival_delay_float'])
data['arrival_delay_float'] = pd.to_timedelta(data['arrival_delay_float']).dt.total_seconds() / 60

data['departure_delay_float'] = data['departure_delay'].astype(str)
data['departure_delay_float'] = data['departure_delay_float'].str.split(' ', n=2).str[-1]
data['departure_delay_float'] = data['departure_delay_float'].astype(str)
data['departure_delay_float'] = pd.to_timedelta(data['departure_delay_float'])
data['departure_delay_float'] = pd.to_timedelta(data['departure_delay_float']).dt.total_seconds() / 60

data.columns

data['actual_arrival_time_float'] = data['actual_arrival_time'].dt.time.astype(str).str.replace(':', '').replace('NaT', '0').astype(int)
data['actual_departure_time_float'] = data['actual_departure_time'].dt.time.astype(str).str.replace(':', '').replace('NaT', '0').astype(int)

data['actual_arrival_time']

data['actual_arrival_time_float']

# create column delay span


data['delay_span'] = '0'

for index in data.index:
    if (data['departure_delay'][index] > pd.Timedelta(0, unit='m')) and (data['departure_delay'][index] <= pd.Timedelta(5, unit='m')):
        data['delay_span'][index] = '0-5'
    elif (data['departure_delay'][index] > pd.Timedelta(5, unit='m')) and (data['departure_delay'][index] <= pd.Timedelta(10, unit='m')):
        data['delay_span'][index] = '5-10'
    elif (data['departure_delay'][index] > pd.Timedelta(10, unit='m')) and (data['departure_delay'][index] <= pd.Timedelta(15, unit='m')):
        data['delay_span'][index] = '10-15'
    elif (data['departure_delay'][index] > pd.Timedelta(15, unit='m')) and (data['departure_delay'][index] <= pd.Timedelta(20, unit='m')):
        data['delay_span'][index] = '15-20'
    elif (data['departure_delay'][index] > pd.Timedelta(20, unit='m')) and (data['departure_delay'][index] <= pd.Timedelta(30, unit='m')):
        data['delay_span'][index] = '20-30'
    elif (data['departure_delay'][index] > pd.Timedelta(30, unit='m')) and (data['departure_delay'][index] <= pd.Timedelta(45, unit='m')):
        data['delay_span'][index] = '30-45'
    elif (data['departure_delay'][index] > pd.Timedelta(45, unit='m')) and (data['departure_delay'][index] <= pd.Timedelta(60, unit='m')):
        data['delay_span'][index] = '45-60'

data['arrival_delay_span'] = '0'

for index in data.index:
    if (data['arrival_delay'][index] > pd.Timedelta(0, unit='m')) and (data['arrival_delay'][index] <= pd.Timedelta(5, unit='m')):
        data['delay_span'][index] = '0-5'
    elif (data['arrival_delay'][index] > pd.Timedelta(5, unit='m')) and (data['arrival_delay'][index] <= pd.Timedelta(10, unit='m')):
        data['delay_span'][index] = '5-10'
    elif (data['arrival_delay'][index] > pd.Timedelta(10, unit='m')) and (data['arrival_delay'][index] <= pd.Timedelta(15, unit='m')):
        data['delay_span'][index] = '10-15'
    elif (data['arrival_delay'][index] > pd.Timedelta(15, unit='m')) and (data['arrival_delay'][index] <= pd.Timedelta(20, unit='m')):
        data['delay_span'][index] = '15-20'
    elif (data['arrival_delay'][index] > pd.Timedelta(20, unit='m')) and (data['arrival_delay'][index] <= pd.Timedelta(30, unit='m')):
        data['delay_span'][index] = '20-30'
    elif (data['arrival_delay'][index] > pd.Timedelta(30, unit='m')) and (data['arrival_delay'][index] <= pd.Timedelta(45, unit='m')):
        data['delay_span'][index] = '30-45'
    elif (data['arrival_delay'][index] > pd.Timedelta(45, unit='m')) and (data['arrival_delay'][index] <= pd.Timedelta(60, unit='m')):
        data['delay_span'][index] = '45-60'

data.columns

data['weather_int'] = 0

for index in data.index:
    if data['current_description'][index] == 'clear sky' or data['current_description'][index] == 'few clouds' or data['current_description'][index] == 'scattered clouds' or data['current_description'][index] == 'broken clouds' or data['current_description'][index] == 'overcast clouds':
        data['weather_int'][index] = 0 # DRY
    elif data['current_description'][index] == 'light rain' or data['current_description'][index] == 'light intensity drizzle' or data['current_description'][index] == 'drizzle' or data['current_description'][index] == 'light intensity drizzle rain' or data['current_description'][index] == 'light intensity shower rain':
        data['weather_int'][index] = 1 # light rain
    elif data['current_description'][index] == 'moderate rain' or data['current_description'][index] == 'heavy intensity rain' or data['current_description'][index] == 'very heavy rain' or data['current_description'][index] == 'thunderstorm with rain' or data['current_description'][index] == 'thunderstorm with heavy rain' or data['current_description'][index] == 'shower rain' or data['current_description'][index] == 'heavy intensity shower rain':
        data['weather_int'][index] = 2 # rain

# create column timespan

data['time_span_arrival'] = '0'
data['time_span_int_arrival'] = 0

for index in data.index:
    if (data['actual_arrival_time'][index].hour >= 6) & (data['actual_arrival_time'][index].hour <= 9):
        data['time_span_arrival'][index] = 'morning rush'
        data['time_span_int_arrival'][index] = 1
    elif (data['actual_arrival_time'][index].hour > 9) & (data['actual_arrival_time'][index].hour < 12):
        data['time_span_arrival'][index] = 'until midday'
        data['time_span_int_arrival'][index] = 2
    elif (data['actual_arrival_time'][index].hour >= 12) & (data['actual_arrival_time'][index].hour <= 14):
        data['time_span_arrival'][index] = 'lunch'
        data['time_span_int_arrival'][index] = 3
    elif (data['actual_arrival_time'][index].hour > 14) & (data['actual_arrival_time'][index].hour < 17):
        data['time_span_arrival'][index] = 'afternoon'
        data['time_span_int_arrival'][index] = 4
    elif (data['actual_arrival_time'][index].hour >= 17) & (data['actual_arrival_time'][index].hour <= 19):
        data['time_span_arrival'][index] = 'evening rush'
        data['time_span_int_arrival'][index] = 5
    elif (data['actual_arrival_time'][index].hour > 19) & (data['actual_arrival_time'][index].hour < 22):
        data['time_span_arrival'][index] = 'evening'
        data['time_span_int_arrival'][index] = 6
    elif (data['actual_arrival_time'][index].hour >= 22) & (data['actual_arrival_time'][index].hour <= 2):
        data['time_span_arrival'][index] = 'night'
        data['time_span_int_arrival'][index] = 7
    elif (data['actual_arrival_time'][index].hour > 2) & (data['actual_arrival_time'][index].hour < 6):
        data['time_span_arrival'][index] = 'early morning'
        data['time_span_int_arrival'][index] = 8

# create column timespan

data['time_span'] = '0'
data['time_span_int'] = 0

for index in data.index:
    if (data['actual_departure_time'][index].hour >= 6) & (data['actual_departure_time'][index].hour <= 9):
        data['time_span'][index] = 'morning rush'
        data['time_span_int'][index] = 1
    elif (data['actual_departure_time'][index].hour > 9) & (data['actual_departure_time'][index].hour < 12):
        data['time_span'][index] = 'until midday'
        data['time_span_int'][index] = 2
    elif (data['actual_departure_time'][index].hour >= 12) & (data['actual_departure_time'][index].hour <= 14):
        data['time_span'][index] = 'lunch'
        data['time_span_int'][index] = 3
    elif (data['actual_departure_time'][index].hour > 14) & (data['actual_departure_time'][index].hour < 17):
        data['time_span'][index] = 'afternoon'
        data['time_span_int'][index] = 4
    elif (data['actual_departure_time'][index].hour >= 17) & (data['actual_departure_time'][index].hour <= 19):
        data['time_span'][index] = 'evening rush'
        data['time_span_int'][index] = 5
    elif (data['actual_departure_time'][index].hour > 19) & (data['actual_departure_time'][index].hour < 22):
        data['time_span'][index] = 'evening'
        data['time_span_int'][index] = 6
    elif (data['actual_departure_time'][index].hour >= 22) & (data['actual_departure_time'][index].hour <= 2):
        data['time_span'][index] = 'night'
        data['time_span_int'][index] = 7
    elif (data['actual_departure_time'][index].hour > 2) & (data['actual_departure_time'][index].hour < 6):
        data['time_span'][index] = 'early morning'
        data['time_span_int'][index] = 8

data['weekend'] = 0

for index in data.index:
    if data['weekday'][index] == 'Monday' or data['weekday'][index] == 'Tuesday' or data['weekday'][index] == 'Wednesday' or data['weekday'][index] == 'Thursday':
        data['weekend'][index] = 0
    if data['weekday'][index] == 'Friday':
        data['weekend'][index] = 1
    if data['weekday'][index] == 'Saturday':
        data['weekend'][index] = 2
    if data['weekday'][index] == 'Sunday':
        data['weekend'][index] = 3

data.to_csv("/Volumes/T7/Master/Processed Data/"+data_month_double+"/20_Final_Data.csv")
